src/test/java/ru/netology/statistic/StatisticsServiceTest.java
package ru.netology.statistic;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

public class StatisticsServiceTest {

    @Test
    void findMax() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 12;

        long actual = service.findMax(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void findMin() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 3;

        long actual = service.findMin(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void findMaxIndex() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 10;

        long actual = service.findMaxIndex(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void findMinIndex() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 5;

        long actual = service.findMinIndex(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void incomeSum() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 85;

        long actual = service.incomeSum(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void averageIncome() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 7;

        long actual = service.averageIncome(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void minIncome() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 6;

        long actual = service.minIncome(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void maxIncome() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 11;

        long actual = service.maxIncome(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void belowAverageIncome() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 5;

        long actual = service.belowAverageIncome(incomesInBillions);

        assertEquals(expected, actual);
    }

    @Test
    void aboveAverageIncome() {
        StatisticsService service = new StatisticsService();

        long[] incomesInBillions = {12, 5, 8, 4, 5, 3, 8, 6, 11, 11, 12};
        long expected = 6;

        long actual = service.aboveAverageIncome(incomesInBillions);

        assertEquals(expected, actual);
    }
}